Change default multi-reduction lowering from innerparallel to innerreduction (#9885)
The 'innerparallel' multi-reduction lowering transposes the innermost
reduction dimension with an outer parallel dimension so that the innestmost
dimension becomes parallel and each vector lane can compute an independent
reduction. The 'innerreduction' multi-reduction lowering keeps the
reduction dimension in its innermost position and generates vector
horizontal reductions as part of the computation. The horizontal
reductions are not generated very efficiently right now but transposing
a 2D vector should always be less efficient than computing a
horizontal reduction. If that is not the case, we should investigate
why.

This PR changes the reduction policy from 'innerparallel' to
'innerreduction'.

This PR is part of a set of changes aimed at improving reductions and
transpositions on RISC-V.