Add multi-stage full build and test (#9877)
This creates jobs to build all of IREE on one machine and then test on
another, passing the build directory around. This same build directory
can be used in other jobs also (e.g. for GPU testing, which I'm punting
on for now).

I ran into issues using the GitHub action artifacts for this. I knew
we'd get suboptimal transfer speeds since it's hosted on Azure and
we'd probably want another solution eventually, but they were much
worse than I expected. So this uses a GCS bucket for the transfer
instead.

We don't want presubmit jobs to be able to affect builds from other
commits, so that's a bit tricky. To get around this, I have separate
buckets for presubmit and postsubmit. The runners only have "create"
and "read" permissions, not "delete" permissions, which means they
cannot remove or overwrite existing artifacts. They upload to a file
including the commit name in its path. If a file is already there, the
upload (and therefore the job) would fail. I think this avoids the
possibility of one build affecting another, even if someone were trying
to do so intentionally. If we move to avoiding rerunning a job if the
artifact for it already exists, we'll need to revisit this.

Part of https://github.com/iree-org/iree/issues/9881