Allow skipping CI with a label (#10055)
This gives a switch for the big binary decision point of "I'm not 
editing code". We determined that we don't have a compelling use case
for picking and choosing which checks run, at least for now. See
https://github.com/iree-org/iree/issues/10042. Note that a status of
"skipped" still counts as "passing" for GitHub's required checks, which
seems wrong to me, but is ok for this use case (otherwise we would
still run the summary and let it compute that everything had been
skipped.

There's a lot of repetition here. Unfortunately, GitHub actions don't
have any sort of variables, not even the standard yaml anchors, and our
best proxy for them, global `env` variables, doesn't work because the
`env` context is not available in the if condition (see
https://docs.github.com/en/actions/learn-github-actions/contexts#context-availability)