Use more extensible method for determining whether to run jobs (#10063)
Using `paths-ignore` as I attempted in
https://github.com/iree-org/iree/pull/10057 doesn't work because we
need the "summary" job to run and post its status. This method will
allow us to implement arbitrary logic for whether to run (including
looking at the paths). If we want to be able to determine *which* jobs
should run dynamically, then we can have this output a JSON list
instead and test whether the job name is in the list, as I demoed in
https://github.com/iree-org/iree/pull/10044. That's a lot of repeated
checks though, so holding off until we need it (and maybe I've
implemented something to generate these files).