Making executable layouts optional for local executables. (#10074)
The inline HAL does not need them and they were always somewhat
optional for tooling like executable_library_benchmark given that
the packing of bindings is well-defined. We still need them for the
full HAL which may have sparse bindings as part of one or more
descriptor sets that are shared across many dispatches.