Use run ID to identify artifacts (#10081)
We currently index by the commit. The idea was that we would be able to
cache artifacts if a commit was rerun. We aren't doing that yet though
and we're having issues because the runners don't have delete
privileges on the objects. This means that a rerun can't overwrite
an artifact created by a previous run.

Run ID + Run attempt is unique. This still works if we rerun a job that
depends on the artifact because they are not computing the artifact
paths themselves, but instead getting it from the creating job's
output. So if the job creating the artifact isn't rerun, the output
will still point to the original artifact.