Add e2e matmul tests on vmvx+ukernels (float32-only for now) (#10193)
Other types than `float32` are blocked on vmvx ukernels support for those (#9903).  I'm interested in landing float32 support early because the path to supporting other data types goes through breaking changes in the existing vmvx ukernel interface for matmul (limiting the generality of the BLAS-inspired interface, particularly the `alpha` and `beta` parameters) so I want to have e2e tests in place at the start of that process.