Change `softmax` test to use `maxf`. (#10219)
The e2e softmax test uses `cmpf` -> `select` for max operations. Use
`maxf`instead. This allows the op to be vectorized. The TOSA to Linalg
lowering has been recently updated to do the same (and this test was
derived from using an older TOSA to Linalg lowering).

Related to PR #10177