Build Linux releases on big managed runners (#10126)
This speed the linux builds up a bit, bringing the time for the longest
job down from ~5 hours to ~20 minutes. Note that this is *only* the
Linux jobs. The mac ones still take about 4 hours. This should still
help when iterating on the release though and for faster failure
indicators (it was indeed helpful when I was iterating here).

I ran into issues when testing because I was using a package suffix in
the workflow dispatch, which evidently had never actually been tested
because it was totally broken. This gave me a lot of wonderful
opportunity to bash my head against bash and I reworked a lot of the
`build_linux_package.sh` script. In retrospect, I wish I'd just removed
the `package_suffix` feature.

Test run: https://github.com/iree-org/iree/actions/runs/2923210349

skip-ci