Optimize tiling sizes heuristics for elementwise dispatches. (#10179)
In the past, small numbers could be picked because we want vectorization
enabled for all the kernels. The PR picks a more reasonable tiling sizes
and addresses tiny dispatch issues.

The peeling pipeline works in IREE and the PR moves elementwise dispatches
(and copy only dispatches) to use peeling approach. In this case, we're
still able to vectorize the dispatches.

This PR changes the logic to limit the unroll factor when computing the
vector level tiling sizes. It avoids generating many operations, which saves
many compilation time and binary size for quantized models. It also
improves models performance that IREE is tracking for all CPU backends.

Fixes https://github.com/iree-org/iree/issues/9660