Pass env variable instructing CI to write Bazel cache (#10248)
So it turns out that the postsubmit jobs were never actually writing to
the remote cache here. How were they still getting cache hits then?
Because until last week, the *Kokoro* postsubmit jobs were still
writing that cache. So given that they were running concurrently with
the GitHub jobs, we had a pattern that was totally consistent with
GitHub actions postsubmit jobs writing the cache: an initial build that
is slow followed by fast builds from a populated cache. The last Kokoro
build of the TF integration binaries was for
https://github.com/iree-org/iree/commit/d04cb91efd, so everything up to
that point was cached. When we next upgraded LLVM and invalidated a big
chunk of the cache, we no longer had anything writing it and builds
became very slow as they got no cache hits. Even more fun, I had ruled
out infrastructure issues because I could revert back to the old commit
and get a fast TF build, but of course that's just because the old
stuff was still cached!

This is basically impossible to test on presubmit because those runners
aren't allowed to write to the cache and I've put a lot of effort into
making it impossible to run presubmits to touch any postsubmit stuff.

(probably) fixes https://github.com/iree-org/iree/issues/10237