Add experimental path to use transformer dialect only in the codegen (#10168)
Add a new option so that we can run transformer dialect after tile and
distribute on workgroup. This allows bypassing the problem of setting
the number of workgroups. For now we just pass a tile size through
command line.
This will then allow writing transform dialect pipelines decided by
KernelConfig. Later the transform dialect can evolve to do the tile and
distribute to workgroup.