Install GitHub runner on startup in a tmpfs (#10318)
The installation is just extracting an archive, so is not costly. This
makes the runner more decoupled from its boot disk. It allows us to
upgrade the runner version (which we'll need to do frequently) without
creating a whole new VM image (which is a pain and requires full VM
replacement to update). The tmpfs also gives us better IO performance,
which should speed a few things along (though we already give Docker a
tmpfs, so the difference won't be huge). It should allow us to switch
to a much simpler base image that just has a few dependencies
installed, rather than the current image, which I accidentally made 1TB
and is the product of much messing around.

Tested: https://github.com/iree-org/iree/pull/10298 targets the testing
runners which are using this config.

skip-ci

This is the last bit of https://github.com/iree-org/iree/pull/10133

Closes https://github.com/iree-org/iree/pull/10133