Starting support for resource attribute serialization. (#10377)
This allows size queries as used in most of the compilation flow to ask
the attributes for their size and verify on attribute types. It's not
currently serializing the contents and will fail, but now we also
properly emit the error instead of crashing.

Instead of assertions as before failures now look like:
```
invalid_constants.mlir:7:3: error: constant attribute failed to serialize: unsupported format or encoding
  vm.rodata private @elided_resource dense_resource<__elided__> : tensor<768x3072xi8>
```

For the special case of elided attributes a new opt-in flag allows zeros
to be written: `--iree-util-zero-fill-elided-attrs`. This should make it
easier to get repros for compilation errors even if at runtime the results will
be wrong.

Fixes #10345.