Adding subspan offset/length to hal.buffer_view.create. (#10378)
It's nearly universal that subspans are required on results as we
allocate the results together. Instead of requiring an allocating
subspan op prior to creating the buffer view this adds the subspan range
to the buffer view creation function. Today this just does the same
subspan+create dance in C but in the future we can optimize that.