Generalize TransposeSharedMem Pipeline to Fusion with Other ElementWise (#10360)
Updates transpose detection and shared mem promotion to support fusion
with non transposed operands.

KernelConfig now identifies transposes based on individual operands that
meet the criteria for shared mem usage then chooses the
TransposeSharedMem pipeline.

LLVMGPUTileAndDistribute only supports a hard coded list of
"operandsToPromote" for promtion filter applied to the entire funcOp.
This is problematic if we want to single out individual operands for
each linalg generic within a funcOp. The solution is to walk the funcOp
for linalg ops and create individual filters specific to promote each
linalg op, then add them all to the pattern and apply.