Make sure GCS dir changes with run attempt (#10410)
Otherwise jobs try to write to the wrong attempt directory. See
https://github.com/iree-org/iree/actions/runs/3049599084/jobs/4926961709
where on attempt 6 it's writing to the directory of attempt 2, which is
when the setup job passed (reruns were only for failures). Would be
really cool if we just had proper variables and didn't have to hack
around with job outputs like this...