[ResolveBufferDescriptors] Fix the handling of rank-reducing subview
Previously the handling of the rank-reducing subview would process only
the first "rank of the subview" strides, sizes, and offset whereas all
the "rank of source" values contribute to the final information.
The only difference when a rank-reducing subview is involved is some of
these values are not replicated in the tuple of results.

E.g., consider the following mask of kept dimensions: [1, 0, 1]
I.e., resDim0 == srcDim0 and resDim1 == srcDim2
Prior to this patch, we would compute:
- For the sizes:
-- resDim0 with srcDim0: good
-- resDim1 with srcDim1: bad
- For the strides:
-- resDim0 with srcDim0: good
-- resDim1 with srcDim1: bad
- For the offset:
-- resOffset: fct(srcDim0, srcDim1): bad

With this patch, we would compute:
- For the sizes:
-- resDim0 with srcDim0: good
-- resDim1 with srcDim2: good
- For the strides:
-- resDim0 with srcDim0: good
-- resDim1 with srcDim2: good
- For the offset:
-- resOffset: fct(srcDim0, srcDim1, srcDim2): good

I believe the previous patch mostly worked, because if you drop only the
dimensions at the end of the shape, your sizes and strides would be
valid (but not your offset, unless I missed something.)