Generalize CI configuration job and enable targetting testing runners (#10623)
This moves these tags to being git trailers
(https://git-scm.com/docs/git-interpret-trailers) rather than a bespoke
system. Note that this means that a "skip-ci" tag must be formatted as
a git trailer with a trailing ':'. The text field afterwards has no
special meaning, but can be used to explain the reason for skipping the
CI.

Tested:
- https://github.com/iree-org/iree/actions/runs/3181833751/ ran on
  testing runners
- https://github.com/iree-org/iree/actions/runs/3175839969/ skipped
  CI