Add a pass to set/unset encoding around compute operations. (#10872)
This adds the pass that sets the encoding around compute operations,
currently only linalg.matmul. Apart from setting the encoding the
pass also introduces necessary padding. The hope is that the pad can
be propogated across operations to amortize the cost of the pad. The
set and unset encoding operations work on the padded sizes.

The follow up PRs will pull this into dispatches appropriately. If
the pad isnt propagated then a pad -> set_encoding (and the
unset_encoding -> slice) operations will get pulled into the same
dispatch allowing for these operations to be folded into the
pack/unpack operations that the set/unset encoding get lowered to
in the backends.