Making status_cc.h an implementation detail of status.h. (#10946)
This avoids the additional library and allows all code that may be
compiling as C++ to access the iree::Status wrapper. Prior to this we
needed the split because of the cc file but by moving to header-only we
avoid the need to ever be able to support a C++ compilation unit.