Rework usage of testing templates (#10972)
This makes it possible to run production templates on testing runners,
which is useful to enable testing of something that has already been
committed and avoids mistakes (that I've made frequently) around the
environment metadata field. It's a bit hacky, but I think will make
updating these significantly easier.

This also names templates with non-prod modifications with a "testing"
suffix. To accomplish this, I had to shorten the template names quite a
bit. I think the result is actually much more readable.

Tested: Deployed new templates to testing runners -> 
https://github.com/iree-org/iree/actions/runs/3364878153/jobs/5579737086

skip-ci: configuration of the runners themselves