Supporting concurrent in-place copies during stream allocation. (#11046)
This will allow us to schedule copies (and other ops) together when
operating on the same resources either when directed by the user at
higher levels of the stack ("trust me this is safe") or by partitioning
that can ensure the ops are mutually exclusive (non-overlapping ranges
derived from analysis). This will be useful for collectives as we can
group multiple collectives together operating on the same send and recv
resources.

Progress on #7729.
Progress on #9580.