Optimize code size of the mmt4d ukernel. (#11026)
Total code size (AArch64, Release) for iree_uk_mmt4d* symbols excluding
tile funcs (`iree_uk_mmt4d_tile*`; the size of these is unaffected by
this PR).

```
Before: 996 bytes
After:  704 bytes
```
So that's a -30% shrink! Now in practice, this code would be used
together with one of the tile funcs, ~300 bytes in average. So the
effective code size shrink (from e.g. a CPU L1 instruction cache
perspective) would be from ~1,300 bytes to ~1,000. Still pretty good.

The over all size of *all* tile funcs is bigger (~2,700 bytes) but only
one of them tends to be used on a given machine, in a given workload.

Part of this is carefully using `__builtin_unreachable`. Another part is
ensuring that all validation is localized in `iree_uk_mmt4d_validate`,
and disabling all that if `NDEBUG`. So there's a little bit of danger
here. There is the possibility of actually reaching a
`__builtin_unreachable` in a Release build (where `NDEBUG` is defined)
if a microkernel is called with bad params.