Create Docker images for the latest software (#11083)
This enables us to test against the latest compiler features and local
dependencies. I'm giving the asan build a little bit of double duty
here in that it will also catch issues specific to later compiler
versions. The potential confusion or extra iteration time from this
combination seems worth it for the compute/latency savings. The asan
build seems a good candidate for using the latest compiler because
sanitizers presumably catch more issues with later versions. We don't
necessarily need to make guarantees that our build works with our
oldest supported compiler version and sanitizers. It seems reasonable
that people running sanitizer builds could update to the latest
versions.

This change also restructures the way we use Docker contexts. We have
already accrued some duplication because of the difficulty factoring
Docker files. Instead, this moves the docker context to include
potentially any file in the repository, which allows reuse via scripts.
We can continue to reduce duplication in future changes.